test: merge leading newline deck test
